requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
playwright>=1.44.0
